+++ src/hooks/purchases/useGetPurchases.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect, useCallback } from "react";
 import { getPurchases } from "@/lib/purchase/getPurchases";
-import { GetPurchaseQueryParams, PurchaseData } from "@/lib/purchase/IPurchase";
+import { GetPurchaseQueryParams, PurchaseData } from "@/lib/purchases/IPurchase";
 import { PaginationResponse } from "@/lib/types";
 import { useQuery } from "../useQuery";
 import { toast } from "@/lib/toast";
